refactor: simplify Autofix presentation boundaries
