+++ packages/control-plane/src/session/message-queue.test.ts
@@ -230,7 +230,7 @@ describe("SessionMessageQueue", () => {
       }),
       event: expect.objectContaining({
         type: "user_message",
-        data: expect.any(String),
+        data: expect.stringContaining('"kind":"review"'),
       }),
       feedbackKey: command.feedbackKey,
       pullRequestKey: "github:99:42",
@@ -247,7 +247,45 @@ describe("SessionMessageQueue", () => {
     expect(h.broadcast).toHaveBeenCalledWith(
       expect.objectContaining({
         type: "sandbox_event",
-        event: expect.objectContaining({ autofix: expect.any(Object) }),
+        event: expect.objectContaining({ origin: command.origin }),
+      })
+    );
+  });
+
+  it("preserves Open Inspect review provenance without projecting it", async () => {
+    const h = buildQueue();
+    const origin = {
+      kind: "open_inspect_review" as const,
+      sourceSessionId: "review-session",
+      sourceMessageId: "review-message",
+      publicationKey: "github-review:opaque",
+      feedbackUrl: "https://github.com/acme/widgets/pull/42#pullrequestreview-1234",
+    };
+
+    await h.queue.enqueueAutofix({
+      type: "enqueue_feedback",
+      feedbackKey: "github:99:review:1234",
+      pullRequest: {
+        repositoryId: "99",
+        number: 42,
+        artifactId: "artifact-1",
+      },
+      prompt: "Address the submitted review feedback.",
+      author: { id: "7", login: "open-inspect[bot]" },
+      origin,
+      attemptLimit: 10,
+    });
+
+    const [admission] = h.repository.admitAutofixMessage.mock.calls[0] as unknown as [
+      Parameters<SessionRepository["admitAutofixMessage"]>[0],
+    ];
+    const event = JSON.parse(String(admission.event.data)) as {
+      origin?: unknown;
+    };
+    expect(event.origin).toEqual(origin);
+    expect(h.broadcast).toHaveBeenCalledWith(
+      expect.objectContaining({
+        event: expect.objectContaining({ origin }),
       })
     );
   });
