+++ packages/shared/src/types/index.ts
@@ -24,6 +24,7 @@ export type {
 
 export {
   githubAutofixEnvelopeSchema,
+  githubAutofixOriginSchema,
   githubAutofixSessionCommandSchema,
   githubAutofixSessionResponseSchema,
 } from "./github-autofix";
