+++ packages/shared/src/types/sandbox-events.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import { sessionDiffBaselineRepositorySchema } from "./session-diffs";
 import { resolvedSessionAttachmentsSchema } from "./session-attachments";
+import { githubAutofixOriginSchema } from "./github-autofix";
 
 const recordSchema = z.record(z.string(), z.unknown());
 const gitSyncStatusSchema = z.enum(["pending", "in_progress", "completed", "failed"]);
@@ -173,14 +174,7 @@ export const sandboxEventSchema = z.discriminatedUnion("type", [
         avatar: z.string().optional(),
       })
       .optional(),
-    autofix: z
-      .object({
-        feedbackKind: z.enum(["pr_comment", "review"]),
-        authorType: z.enum(["human", "bot"]),
-        pullRequestNumber: z.number().int().positive(),
-        feedbackUrl: z.url(),
-      })
-      .optional(),
+    origin: githubAutofixOriginSchema.optional(),
     // Attachment metadata only — never inline content, which would bloat the
     // events table and every broadcast. attachmentId lets clients stream attachments.
     attachments: resolvedSessionAttachmentsSchema.optional(),
