+++ packages/web/src/components/session-timeline.tsx
@@ -462,16 +462,17 @@ function UserMessageEvent({
       copyButtonClassName="p-1 text-secondary-foreground hover:text-foreground hover:bg-muted/60 opacity-0 pointer-events-none group-hover:opacity-100 group-hover:pointer-events-auto focus-visible:opacity-100 focus-visible:pointer-events-auto transition-colors"
       onCopyContent={onCopyContent}
     >
-      {event.autofix && (
+      {event.origin && (
         <div className="mb-3 flex flex-wrap items-center gap-x-2 gap-y-1 border-b border-border pb-2 text-xs">
           <span className="font-medium text-accent">Resumed by PR feedback</span>
           <span className="text-muted-foreground">
-            {event.autofix.feedbackKind === "review" ? "Review" : "PR comment"} ·{" "}
-            {event.autofix.authorType === "bot" ? "Bot" : "Human"} · PR #
-            {event.autofix.pullRequestNumber}
+            {event.origin.kind === "pr_comment" ? "PR comment" : "Review"} ·{" "}
+            {event.origin.kind === "open_inspect_review" || event.origin.authorType === "bot"
+              ? "Bot"
+              : "Human"}
           </span>
           <a
-            href={event.autofix.feedbackUrl}
+            href={event.origin.feedbackUrl}
             target="_blank"
             rel="noreferrer"
             className="text-accent hover:underline"
